requirements.txt
streamlit
opencv-python
deepface
python-dotenv
matplotlib
pandas 